pyproject.toml
[tool.poetry]
name = "nonebot_plugin_ayaka"
version = "0.1.0"
description = "a useful plugin providing convinient tools for the development of textual game on QQ"
authors = ["Su <wxlxy316@163.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.0.0b5"
nonebot-adapter-onebot = "^2.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
